% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_smoothed.R
\name{lda_smoothed}
\alias{lda_smoothed}
\title{Run LDA adapted to use a count matrix}
\usage{
lda_smoothed(
  N,
  K,
  max_iter = 50,
  thresh = NULL,
  seed = NULL,
  cores = NULL,
  alpha = NULL,
  eta = NULL,
  NMF = FALSE
)
}
\arguments{
\item{N}{matrix of word counts}

\item{K}{the number of topics to look for}

\item{max_iter}{the maximum number of EM iterations to run}

\item{thresh}{if you want to set a specific threshold for L convergence,
if NULL it's defined as 0.01 percent of the previous value}

\item{seed}{if you want a reproducible result you can set a seed,
if NULL the topic distributions are initialised from K random documents,}

\item{cores}{number of cores to run the E-step in parallel,
if NULL all detected cores are used}

\item{alpha}{if you want to set the exchangeable Dirichlet parameter for theta,
if NULL a default value of 1/K is used}

\item{eta}{if you want to set the exchangeable Dirichlet parameter for beta,
if NULL a default value of 1/K is used}

\item{NMF}{logical indicating if lambda should be initialised using non-negative matrix factorisation,
if FALSE it is generated using K random documents}
}
\description{
Run LDA adapted to use a count matrix
}
