% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_original.R, R/LDA_original_par.R,
%   R/LDA_noalpha.R
\name{lda_original}
\alias{lda_original}
\alias{lda_original_par}
\alias{lda_noalpha}
\title{Run LDA as in the Blei 2003 paper}
\usage{
lda_original(docs, K, max_iter = 50, thresh = NULL, seed = NULL)

lda_original_par(
  docs,
  K,
  max_iter = 50,
  thresh = NULL,
  seed = NULL,
  cores = NULL
)

lda_noalpha(
  docs,
  K,
  max_iter = 50,
  thresh = NULL,
  seed = NULL,
  cores = NULL,
  alpha = NULL
)
}
\arguments{
\item{docs}{a list containing all the documents, with the vocabulary encoded
e.g. docs[[1]] = c(1, 5, 2)  would represent the word indices from a pre-defined vocabulary}

\item{K}{the number of topics to look for}

\item{max_iter}{the maximum number of EM iterations to run}

\item{thresh}{if you want to set a specific threshold for L convergence,
if NULL it's defined as 0.01 percent of the previous value}

\item{seed}{if you want a reproducible result you can set a seed,
if NULL the topic distributions are initialised from K random documents,}

\item{cores}{number of cores to run the E-step in parallel,
if NULL all detected cores are used}

\item{alpha}{if you want to set the exchangeable Dirichlet parameter for theta,
if NULL a default value of 1/K is used}
}
\value{
A list of all parameters
}
\description{
Run LDA as in the Blei 2003 paper
}
\section{Functions}{
\itemize{
\item \code{lda_original_par}: Runs E-step in parallel

\item \code{lda_noalpha}: Alpha is fixed
}}

