% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_smoothed.R
\name{lda_smoothed}
\alias{lda_smoothed}
\title{Run LDA adapted to use a count matrix}
\usage{
lda_smoothed(
  N,
  K,
  max_iter = 50,
  thresh = 1e-04,
  seed = NULL,
  cores = NULL,
  alpha = NULL,
  eta = NULL,
  NMF = FALSE
)
}
\arguments{
\item{N}{matrix of word counts}

\item{K}{the number of topics to look for}

\item{max_iter}{the maximum number of EM iterations to run}

\item{thresh}{threshold for L convergence, (L_i - L_{i-1})/L_i < thresh}

\item{seed}{set a seed for the random documents to initialise beta}

\item{cores}{number of cores to run the E-step in parallel,
if NULL all detected cores are used}

\item{alpha}{if you want to set the exchangeable Dirichlet parameter for theta,
if NULL a default value of 1/K is used}

\item{eta}{the exchangeable Dirichlet parameter for beta, if NULL a default value of 1/K is used}

\item{NMF}{logical indicating if lambda should be initialised using non-negative matrix factorisation,
if FALSE it is generated using K random documents}
}
\description{
Run LDA adapted to use a count matrix
}
